pyproject.toml
[project]
name = "new"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.135.2",
    "google-generativeai>=0.8.6",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.22",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langchain-openai>=1.1.12",
    "python-dotenv>=1.2.2",
    "sympy>=1.14.0",
    "uvicorn>=0.42.0",
]
